requirements.txt
torchvision
matplotlib
pillow
matplotlib